Refactor: 각 ui마다 폴더 생성 및 style 코드 파일 따로 생성
