+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "joseph-ui-kit",
   "description": "ui-kit that make user convenient",
-  "version": "0.2.4",
+  "version": "0.2.5",
   "private": false,
   "main": "dist/index.js",
   "module": "dist/index.js",
