+++ src/components/Accordion/Accordion.stories.mdx
@@ -39,7 +39,7 @@ export const Template = (args) => <Accordion {...args} />;
 
 <Canvas>
   <Story
-    name="Primary"
+    name="Default"
     args={{
       title: "title is here.",
       children: (
@@ -58,4 +58,4 @@ export const Template = (args) => <Accordion {...args} />;
 
 ## Component Api
 
-<ArgsTable story="Primary" />
+<ArgsTable story="Default" />
