+++ src/components/Accordion/Accordion.tsx
@@ -1,0 +1,37 @@
+/** @jsxImportSource @emotion/react */
+import { css } from "@emotion/react";
+import * as style from "./Accordion.style";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faAngleUp, faAngleDown } from "@fortawesome/free-solid-svg-icons";
+import React, { useState } from "react";
+
+type AccordionProps = {
+  title: string;
+  children: JSX.Element;
+};
+
+const Accordion = ({ title, children }: AccordionProps) => {
+  const [open, setOpen] = useState(false);
+
+  const handleAccordion = () => {
+    setOpen((open) => !open);
+  };
+
+  return (
+    <div css={style.container}>
+      <button tabIndex={-1} css={style.button(open)} onClick={handleAccordion}>
+        {title}
+        <div css={style.arrowIcon}>
+          {open ? (
+            <FontAwesomeIcon icon={faAngleUp} />
+          ) : (
+            <FontAwesomeIcon icon={faAngleDown} />
+          )}
+        </div>
+      </button>
+      {open ? <div css={style.content(open)}>{children}</div> : null}
+    </div>
+  );
+};
+
+export default Accordion;
