+++ src/components/Button/Button.stories.mdx
@@ -115,7 +115,7 @@ export const Template = (args) => <Button {...args} />;
 
 <Canvas>
   <Story
-    name="Primary"
+    name="Default"
     args={{
       bgColor: "#0e62fe",
       hover: {
@@ -136,7 +136,7 @@ export const Template = (args) => <Button {...args} />;
 
 ## Component Api
 
-<ArgsTable story="Primary" />
+<ArgsTable story="Default" />
 
 <Canvas>
   <Story
