+++ src/components/DropDown/DropDown.stories.mdx
@@ -72,7 +72,7 @@ export const Template = (args) => <DropDown {...args} />;
 
 <Canvas>
   <Story
-    name="Primary"
+    name="Default"
     args={{
       items: [
         { id: "option1", text: "text1", disabled: true },
@@ -91,4 +91,4 @@ export const Template = (args) => <DropDown {...args} />;
 
 ## Component Api
 
-<ArgsTable story="Primary" />
+<ArgsTable story="Default" />
