+++ src/components/DropDown/DropDown.tsx
@@ -1,59 +1,9 @@
 /** @jsxImportSource @emotion/react */
 import { css } from "@emotion/react";
 import React, { useState, useRef, useLayoutEffect } from "react";
+import * as style from "./DropDown.style";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faAngleUp, faAngleDown } from "@fortawesome/free-solid-svg-icons";
-
-const container = css`
-  position: relative;
-  display: inline-block;
-`;
-
-const titleAndWarnText = css`
-  margin: 5px 0px;
-  font-size: 12px;
-  color: gray;
-`;
-
-const button = (open: boolean, size: string) => css`
-  width: 400px;
-  height: ${size === "small" ? "32px" : size === "middle" ? "40px" : "48px"};
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0 15px;
-  border: none;
-  border-bottom: 1px solid #d1d1d1;
-  background-color: #f4f4f4;
-  font-size: 14px;
-  color: #161616;
-  box-sizing: border-box;
-  :hover {
-    background-color: ${open ? "none" : "#e8e8e8"};
-  }
-  :focus {
-    outline: ${open ? "none" : "2px solid #0f62fe"};
-  }
-`;
-
-const arrowIcon = css`
-  width: 10%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-const suggestion = (open: boolean, zIndex: number) => css`
-  position: absolute;
-  width: 400px;
-  height: 200px;
-  background-color: #f4f4f4;
-  overflow-x: none;
-  overflow-y: auto;
-  outline: ${open ? "1px solid #0f62fe" : "none"};
-  z-index: ${zIndex};
-`;
 
 const suggestionItem = (
   index: number,
@@ -142,9 +92,9 @@ const DropDown = ({
   });
 
   return (
-    <div css={container}>
+    <div css={style.container}>
       {hideLabel ? null : (
-        <label htmlFor="dropdown" css={titleAndWarnText}>
+        <label htmlFor="dropdown" css={style.titleAndWarnText}>
           {label}
         </label>
       )}
@@ -152,12 +102,12 @@ const DropDown = ({
         id="dropdown"
         ref={buttonRef}
         tabIndex={-1}
-        css={button(open, size)}
+        css={style.button(open, size)}
         onBlur={closeSuggestion}
         onClick={handleSuggestion}
       >
         <span>{items[seletedItem].text}</span>
-        <div css={arrowIcon}>
+        <div css={style.arrowIcon}>
           {open ? (
             <FontAwesomeIcon icon={faAngleUp} />
           ) : (
@@ -167,7 +117,7 @@ const DropDown = ({
       </button>
 
       {open ? (
-        <div ref={suggestionRef} css={suggestion(open, zIndex)}>
+        <div ref={suggestionRef} css={style.suggestion(open, zIndex)}>
           {items.map((item, index) => (
             <div
               key={item.id}
@@ -183,7 +133,7 @@ const DropDown = ({
           ))}
         </div>
       ) : null}
-      {hideWarn ? null : <div css={titleAndWarnText}>{warn}</div>}
+      {hideWarn ? null : <div css={style.titleAndWarnText}>{warn}</div>}
     </div>
   );
 };
