+++ src/components/Loading/Loading.stories.mdx
@@ -39,11 +39,11 @@ import Loading from "./Loading";
 export const Template = (args) => <Loading {...args} />;
 
 <Canvas>
-  <Story name="Primary" args={{}}>
+  <Story name="Default" args={{}}>
     {Template.bind({})}
   </Story>
 </Canvas>
 
 ## Component Api
 
-<ArgsTable story="Primary" />
+<ArgsTable story="Default" />
