+++ src/components/Loading/Loading.style.tsx
@@ -1,8 +1,7 @@
 /** @jsxImportSource @emotion/react */
 import { css, keyframes } from "@emotion/react";
-import React from "react";
 
-const circleAnimation = keyframes`
+export const circleAnimation = keyframes`
   from {
     transform: translate(0%, 0%) rotate(0);
   }
@@ -10,7 +9,7 @@ const circleAnimation = keyframes`
     transform: translate(0%, 0%) rotate(360deg);
   }
 `;
-const circle = (loadingSize: number) => css`
+export const circle = (loadingSize: number) => css`
   width: ${loadingSize}px;
   height: ${loadingSize}px;
   border: ${loadingSize / 8}px solid #0e62fe;
@@ -22,24 +21,3 @@ const circle = (loadingSize: number) => css`
   animation-iteration-count: infinite;
   animation-timing-function: linear;
 `;
-
-type LoadingProps = {
-  size: string;
-  active: boolean;
-};
-
-const Loading = ({ size, active }: LoadingProps) => {
-  let loadingSize = 40;
-
-  if (size === "small") {
-    loadingSize = 40;
-  } else if (size === "middle") {
-    loadingSize = 48;
-  } else if (size === "large") {
-    loadingSize = 56;
-  }
-
-  return active ? <div css={circle(loadingSize)} /> : null;
-};
-
-export default Loading;
