+++ src/components/Modal/Modal.stories.mdx
@@ -114,7 +114,7 @@ export const Template = (args) => <Modal {...args} />;
 
 <Canvas>
   <Story
-    name="Primary"
+    name="Default"
     args={{
       title: "Modal title is here",
       children: (
@@ -138,4 +138,4 @@ export const Template = (args) => <Modal {...args} />;
 
 ## Component Api
 
-<ArgsTable story="Primary" />
+<ArgsTable story="Default" />
