+++ src/components/Search/Search.stories.mdx
@@ -40,9 +40,9 @@ import Search from "./Search";
 export const Template = (args) => <Search {...args} />;
 
 <Canvas>
-  <Story name="Primary">{Template.bind({})}</Story>
+  <Story name="Default">{Template.bind({})}</Story>
 </Canvas>
 
 ## Component Api
 
-<ArgsTable story="Primary" />
+<ArgsTable story="Default" />
