+++ src/components/Tabs/Tabs.stories.mdx
@@ -37,7 +37,7 @@ export const Template = (args) => <Tabs {...args} />;
 
 <Canvas>
   <Story
-    name="Primary"
+    name="Default"
     args={{
       width: "600px",
       height: "50px",
@@ -73,4 +73,4 @@ export const Template = (args) => <Tabs {...args} />;
 
 ## Component Api
 
-<ArgsTable story="Primary" />
+<ArgsTable story="Default" />
